Fixes websocket masking problem on server side and make examples work (#42)
* Fixes websocket masking problem on server side

The client establishes a websocket connection and disconnects when receiving the message from server for the first time. And send the error message to server: A server must not mask any frames that it sends to the client. The server-side error is shown below:
[Error] Error occurred handling client messages.
socket closed

The reason is that the data sent by websocket server cannot be masked. Ormin websocket client object referenced AsyncSocket which proc about websocket is discarded. Ormin websocket client object should update the reference to the new object AsyncWebSocket.

* Fixes examples building